Add logging to help users with encoding issues
We still haven't been able to get ontop of encoding issues in this app.
Most seem to happen when piping from stdout for which I am not clear
on the reasons. The logging information in this script may provide a
helpful prompt but a more permanent solution would be good.